requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
Markdown==3.0.1
psycopg2-binary==2.7.5
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
whitenoise==3.3.0
pdfplumber==0.5.11
